LU-12345 mdd: read, update update mc_flags field atomically
The `mc_flags` field of `struct mdd_changelog` is currently protected
by `mc_lock`, but only for updates -- not for reads.

While this is unlikely to cause any harm, it is technically a data
race.

The data race can be resolved by making `mc_flags` an `atomic_t` and
using `atomic_or()` and `atomic_read()`.

An additional benefit of this change is that it is slightly more
efficient than acquiring a spinlock, and means that the spinlock
protects fewer fields as well.

Signed-off-by: Thomas Bertschinger <bertschinger@lanl.gov>
Change-Id: I43c40aebec6c6194c9906e40bba407198549288e